add: introduce Docker support with multi-stage builds and `docker-compose.yml` remove: eliminate Appveyor, Jenkins, Travis CI configurations and related scripts update: update `spdlog` to v1.17.0 refactor: replace outdated logging and state debugging with safer type casts refactor: restructure database schema into `schema/` directory cleanup: remove unused `COTIRE` references from build system

+++ src/char/src/party.cpp
@@ -293,7 +293,7 @@ void party_reply(const RoseCommon::Packet::CliPartyReply& packet, CCharServer& s
             server.add_user_to_party(user, other->get_party());
             break;
         default:
-            logger->debug("{} replied {} to {}", user.get_name(), packet.get_type(), other->get_name());
+            logger->debug("{} replied {} to {}", user.get_name(), static_cast<uint8_t>(packet.get_type()), other->get_name());
             break;
     }
 }
